pyproject.toml
[tool.poetry]
name = "mbo-bulk-client-import"
version = "0.1.0"
description = ""
authors = ["Rob Marshall <rob@robmarshall.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
pandas = "^2.2.3"
selenium = "^4.27.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
